requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.4.0
PyMuPDF==1.23.8
pdf2image==1.16.3
numpy==1.24.3
easyocr==1.7.0
streamlit==1.28.1
matplotlib==3.7.2
google-generativeai>=0.3.0
python-dotenv>=1.0.0
python-dotenv>=1.0.0
